requirements.txt
alembic==1.2.1
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.2
packaging==19.2
pandas==0.25.1
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.0
pytest-flask==0.15.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
